+++ CHANGE_LOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add org.starchartlabs.alloy.core.collections.MoreSpliterators for streamlined and reduced boilerplate in relation to Java spliterator constructs
 - Addition of support for short-circuited spliterators via org.starchartlabs.alloy.core.collections.MoreSpliterators.shortCircuit
+- Addition of support for paged-data-based spliterators via org.starchartlabs.alloy.core.collections.MoreSpliterators.ofPaged and org.starchartlabs.alloy.core.collections.PageProvider
 
 ## [0.3.0]
 ### Added